Add exceptions to TyreServiceImpl

+++ StisEJB-ejb/src/main/java/cz/muni/fi/pa165/stis/service/impl/TyreServiceImpl.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package cz.muni.fi.pa165.stis.service.impl;
 
 import cz.muni.fi.pa165.stis.dao.TyreDAOLocal;
@@ -23,21 +19,43 @@ public class TyreServiceImpl implements TyreServiceLocal {
 
     @Override
     public void create(Tyre tyre) {
+        if (tyre == null) {
+            throw new IllegalArgumentException("tyre is null");
+        }
+        
+        if (tyre.getId() != null) {
+            throw new IllegalArgumentException("tyre.id is not null");
+        }
         dao.create(tyre);
     }
 
     @Override
     public Tyre get(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Tyre.Id is null");
+        }
         return dao.get(id);
     }
 
     @Override
     public void update(Tyre tyre) {
+        if (tyre == null) {
+            throw new IllegalArgumentException("Tyre is null");
+        }        
+        if (tyre.getId() == null) {
+            throw new IllegalArgumentException("Tyre.id is null");
+        }
         dao.update(tyre);
     }
 
     @Override
     public void remove(Tyre tyre) {
+        if (tyre == null) {
+            throw new IllegalArgumentException("Tyre is null");
+        }        
+        if (tyre.getId() == null) {
+            throw new IllegalArgumentException("Tyre.id is null");
+        }
         dao.remove(tyre);
     }
 
@@ -48,6 +66,9 @@ public class TyreServiceImpl implements TyreServiceLocal {
 
     @Override
     public List<Tyre> findByName(String name) {
+        if (name == null) {
+            throw new IllegalArgumentException("null name");
+        }
         return dao.findByName(name);
     }
